requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.6.2
click==8.1.7
coverage==7.5.3
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy==1.10.0
mypy-extensions==1.0.0
orjson==3.10.3
packaging==24.0
pluggy==1.5.0
pydantic==2.7.2
pydantic-settings==2.2.1
pydantic_core==2.18.3
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
ruff==0.4.7
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
